change for available on x32
on x32 program doesn't work

+++ source/core/typesofoptimizevar.cpp
@@ -39,7 +39,8 @@ void TypesOfOptimizeVar::setRandomAvValue()
 //            }
 //        }
 
-        av_i = rand.global()->bounded(0, desc_link.size());
+//        av_i = rand.global()->bounded(0, desc_link.size());
+        av_i = rand.global()->bounded((int)0, (int)desc_link.size());
         av = desc_link[av_i];
 
         iv_i = av_i;
